src/requirements.txt
Flask
Flask-Admin
Flask-SQLAlchemy
Flask-Security>=1.7.5
setuptools
numpy
pandas